feat(cache): add env files to buildDependencies
