+++ e2e/cases/cache/build-dependencies/index.test.ts
@@ -5,48 +5,6 @@ import { expect, test } from '@playwright/test';
 import type { RsbuildConfig } from '@rsbuild/core';
 import fse from 'fs-extra';
 
-test('`buildCache.cacheDigest` should work as expected', async () => {
-  const cacheDirectory = path.resolve(
-    import.meta.dirname,
-    './node_modules/.cache/test-cache-digest',
-  );
-
-  fs.rmSync(cacheDirectory, { recursive: true, force: true });
-
-  const getBuildConfig = (input: string) => ({
-    cwd: import.meta.dirname,
-    rsbuildConfig: {
-      tools: {
-        bundlerChain: (chain) => {
-          if (input === 'foo') {
-            chain.resolve.extensions.prepend('.test.js');
-          }
-        },
-      },
-      performance: {
-        buildCache: {
-          cacheDirectory,
-          cacheDigest: [input],
-        },
-      },
-    } as RsbuildConfig,
-  });
-
-  // first build without cache
-  let rsbuild = await build(getBuildConfig(''));
-
-  expect(
-    (await rsbuild.getIndexFile()).content.includes('222222'),
-  ).toBeTruthy();
-
-  rsbuild = await build(getBuildConfig('foo'));
-
-  // extension '.test.js' should work
-  expect(
-    (await rsbuild.getIndexFile()).content.includes('111111'),
-  ).toBeTruthy();
-});
-
 test('`buildCache.buildDependencies` should work as expected', async () => {
   const cacheDirectory = path.resolve(
     import.meta.dirname,
