+++ e2e/cases/cache/build-dependencies/src/index.js
