+++ e2e/cases/cache/build-dependencies/tsconfig.json
