+++ e2e/cases/cache/cache-digest/src/a.js
@@ -1,0 +1,1 @@
+export const a = '222222';
