+++ e2e/cases/cache/cache-digest/src/a.test.js
@@ -1,0 +1,1 @@
+export const a = '111111';
