+++ e2e/cases/cache/cache-digest/src/index.js
@@ -1,0 +1,3 @@
+import { a } from './a';
+
+console.log('foo', a);
